requirements.txt
Django==4.2.2
django_environ==0.10.0
Requests==2.31.0
